feat(addons): extensibilidade via addons com o fiscal como primeiro addon nativo
Arquitetura microkernel (padrão VS Code/Medusa/Payload): core mínimo +
contribution points declarativos + contratos estáveis + DI + isolamento
de falhas. Qualquer um estende criando um pacote com defineAddon() e
registrando em 1 linha no host.

- @finopenpos/addon-kit: SDK do microkernel — AddonDefinition (manifest,
  routers, tables, subscriptions, messages, setup, seed), AddonContext
  injetado (db, eventStore, Host API v1, logger), registry com validação
  de colisões, dispatcher de eventos com isolamento de falhas, deep-merge
  de i18n; lado UI com AddonUIDefinition (nav/pages), matchAddonPage e
  AddonServicesProvider/useAddonTRPC (cliente tRPC tipado por recorte)
- @finopenpos/addon-fiscal: todo o domínio fiscal movido para fora do
  core — schema (cities, fiscal_settings, invoices, itens, eventos),
  routers (fiscal, fiscalSettings, cities), serviços SEFAZ, seed IBGE,
  mensagens i18n e as 3 páginas admin; zero imports de internals do host
  (verificado: nenhum @/ no pacote)
- event-sourcing: hook onAppended no EventStore — assinaturas de addons
  reagem a eventos de domínio sem acoplar o core
- host: registro em lib/addons (installed/installed-ui/host-api/
  bootstrap), rota catch-all /admin/[...addon] montando páginas
  contribuídas, nav do admin extensível, i18n mesclado por addon,
  schema drizzle composto, seed e testes via contribution points
- contrato preservado: mesmas rotas tRPC/OpenAPI e URLs /admin/fiscal/*
- testes: 86 web + 10 addon-kit + 5 event-sourcing, todos verdes;
  next build ok; ADR em docs/13-addons.md
