+++ apps/web/src/lib/addons/schema.ts
@@ -1,0 +1,5 @@
+/**
+ * Composição de schema dos addons instalados — re-exportado pelo schema do
+ * app para que o drizzle-kit gere/aplique as tabelas contribuídas.
+ */
+export * from "@finopenpos/addon-fiscal/schema";
