+++ apps/web/src/lib/db/index.ts
@@ -3,11 +3,10 @@ import { drizzle } from "drizzle-orm/pglite";
 import * as schema from "./schema";
 
 const globalForPGlite = globalThis as unknown as {
-  pglite: PGlite | undefined;
+	pglite: PGlite | undefined;
 };
 
-export const pglite =
-  globalForPGlite.pglite ?? new PGlite("./data/pglite");
+export const pglite = globalForPGlite.pglite ?? new PGlite("./data/pglite");
 
 globalForPGlite.pglite = pglite;
