+++ apps/web/src/lib/db/schema.ts
@@ -1,1 +1,3 @@
 export * from "@finopenpos/db/schema";
+// Tabelas contribuídas pelos addons instalados (composição de schema)
+export * from "@/lib/addons/schema";
