+++ apps/web/src/lib/db/seed.ts
@@ -11,7 +11,6 @@ import {
 } from "@/lib/es";
 import { auth } from "../auth";
 import { db } from ".";
-import { cities } from "./schema";
 
 const DEMO_EMAIL = "test@example.com";
 const DEMO_PASSWORD = "test1234";
@@ -275,86 +274,9 @@ export async function seed() {
 		});
 	}
 
-	// ── Cities (IBGE) ─────────────────────────────────────────────────────────
-	const cityCount = await seedCities();
-
 	console.log(
 		`Seeded (event log): 3 payment methods, 1 demo user (${DEMO_EMAIL} / ${DEMO_PASSWORD}), ` +
 			`20 customers, ${seededProducts.length} products, ` +
-			`${orderCount} orders, ${expenseCount} expense transactions, ` +
-			`${cityCount} cities`,
+			`${orderCount} orders, ${expenseCount} expense transactions`,
 	);
-}
-
-const STATES = [
-	"AC",
-	"AL",
-	"AM",
-	"AP",
-	"BA",
-	"CE",
-	"DF",
-	"ES",
-	"GO",
-	"MA",
-	"MG",
-	"MS",
-	"MT",
-	"PA",
-	"PB",
-	"PE",
-	"PI",
-	"PR",
-	"RJ",
-	"RN",
-	"RO",
-	"RR",
-	"RS",
-	"SC",
-	"SE",
-	"SP",
-	"TO",
-];
-
-async function seedCities(): Promise<number> {
-	const existingCities = await db
-		.select({ count: sql<number>`count(*)` })
-		.from(cities);
-
-	if (existingCities[0].count > 0) return existingCities[0].count;
-
-	let total = 0;
-
-	for (const uf of STATES) {
-		try {
-			const res = await fetch(
-				`https://servicodados.ibge.gov.br/api/v1/localidades/estados/${uf}/municipios`,
-			);
-			if (!res.ok) {
-				console.warn(`Failed to fetch cities for ${uf}: ${res.status}`);
-				continue;
-			}
-
-			const data: Array<{ id: number; nome: string }> = await res.json();
-
-			if (data.length > 0) {
-				// Insert in batches of 500 to avoid query size limits
-				for (let i = 0; i < data.length; i += 500) {
-					const batch = data.slice(i, i + 500);
-					await db.insert(cities).values(
-						batch.map((city) => ({
-							id: city.id,
-							name: city.nome,
-							state_code: uf,
-						})),
-					);
-				}
-				total += data.length;
-			}
-		} catch (err) {
-			console.warn(`Error fetching cities for ${uf}:`, err);
-		}
-	}
-
-	return total;
 }
