+++ apps/web/src/messages/en.ts
@@ -45,9 +45,6 @@ const messages = {
     orders: "Orders",
     paymentMethods: "Payment Methods",
     pos: "Point of Sale",
-    fiscal: "Fiscal",
-    fiscalSettings: "Fiscal Settings",
-    invoices: "Invoices",
     adminPanel: "Admin Panel",
     openMenu: "Open menu",
     myAccount: "My Account",
@@ -225,114 +222,6 @@ const messages = {
     outOfStock: "{name} is out of stock",
     limitedStock: "Only {count} units of {name} available",
     limitedUnits: "Only {count} units available",
-  },
-  fiscal: {
-    title: "Invoices",
-    subtitle: "Electronic invoices (NF-e / NFC-e)",
-    noInvoices: "No invoices found",
-    emitInvoice: "Emit Invoice",
-    emitNfce: "Emit NFC-e",
-    emitNfe: "Emit NF-e",
-    syncPending: "Sync Pending",
-    syncResult: "Synced {authorized} of {total} invoices",
-    model: "Model",
-    series: "Series",
-    number: "Number",
-    accessKey: "Access Key",
-    recipient: "Recipient",
-    issuedAt: "Issued At",
-    authorizedAt: "Authorized At",
-    statusCode: "Status Code",
-    statusMessage: "Status Message",
-    protocolNumber: "Protocol",
-    contingency: "Contingency",
-    contingencyType: "Contingency Type",
-    contingencyReason: "Reason",
-    downloadXml: "Download XML",
-    cancelInvoice: "Cancel Invoice",
-    cancelReason: "Cancellation reason (min. 15 characters)",
-    cancelSuccess: "Invoice cancelled successfully",
-    cancelError: "Failed to cancel invoice",
-    voidRange: "Void Number Range",
-    voidReason: "Voiding reason (min. 15 characters)",
-    startNumber: "Start Number",
-    endNumber: "End Number",
-    voidSuccess: "Numbers voided successfully",
-    voidError: "Failed to void numbers",
-    emitSuccess: "Invoice emitted successfully",
-    emitError: "Failed to emit invoice",
-    emitContingency: "Invoice emitted in contingency mode",
-    details: "Invoice Details",
-    items: "Items",
-    events: "Events",
-    xmlViewer: "XML",
-    authorized: "Authorized",
-    rejected: "Rejected",
-    cancelled: "Cancelled",
-    denied: "Denied",
-    pending: "Pending",
-    voided: "Voided",
-    nfe: "NF-e",
-    nfce: "NFC-e",
-    eventCancellation: "Cancellation",
-    eventVoiding: "Voiding",
-    sefazOnline: "SEFAZ Online",
-    sefazOffline: "SEFAZ Offline",
-    toggleNfce: "Emit NFC-e",
-    searchPlaceholder: "Search invoices...",
-    filterByStatus: "Status",
-    filterByModel: "Model",
-  },
-  fiscalSettings: {
-    title: "Fiscal Settings",
-    subtitle: "Configure your company fiscal data and digital certificate",
-    companyInfo: "Company Information",
-    companyName: "Company Name (Legal)",
-    tradeName: "Trade Name",
-    taxId: "CNPJ",
-    stateTaxId: "State Tax ID (IE)",
-    taxRegime: "Tax Regime",
-    taxRegimeSimples: "Simples Nacional",
-    taxRegimeSimplesExcess: "Simples Nacional (excess)",
-    taxRegimeNormal: "Normal (Lucro Presumido/Real)",
-    address: "Address",
-    stateCode: "State (UF)",
-    cityCode: "City Code (IBGE)",
-    cityName: "City Name",
-    street: "Street",
-    streetNumber: "Number",
-    district: "District",
-    zipCode: "ZIP Code (CEP)",
-    addressComplement: "Complement",
-    environment: "Environment",
-    production: "Production",
-    homologation: "Homologation (Testing)",
-    numbering: "Numbering",
-    nfeSeries: "NF-e Series",
-    nfceSeries: "NFC-e Series",
-    nextNfeNumber: "Next NF-e Number",
-    nextNfceNumber: "Next NFC-e Number",
-    nfceSettings: "NFC-e Settings",
-    cscId: "CSC ID",
-    cscToken: "CSC Token",
-    certificate: "Digital Certificate",
-    certificateFile: "Certificate File (.pfx)",
-    certificatePassword: "Certificate Password",
-    certificateUploaded: "Certificate loaded",
-    certificateNotUploaded: "No certificate uploaded",
-    certificateExpiry: "Valid until",
-    certificateCommonName: "Issued to",
-    testConnection: "Test Connection",
-    testSuccess: "SEFAZ connection successful",
-    testError: "SEFAZ connection failed",
-    defaults: "Default Fiscal Fields",
-    defaultNcm: "Default NCM",
-    defaultCfop: "Default CFOP",
-    defaultIcmsCst: "Default ICMS CST",
-    defaultPisCst: "Default PIS CST",
-    defaultCofinsCst: "Default COFINS CST",
-    saved: "Fiscal settings saved",
-    saveError: "Failed to save fiscal settings",
   },
   error: {
     somethingWentWrong: "Sorry, something went wrong",
