+++ apps/web/src/messages/pt-BR.ts
@@ -47,9 +47,6 @@ const messages: Messages = {
     orders: "Pedidos",
     paymentMethods: "Formas de Pagamento",
     pos: "Ponto de Venda",
-    fiscal: "Fiscal",
-    fiscalSettings: "Configurações Fiscais",
-    invoices: "Notas Fiscais",
     adminPanel: "Painel Admin",
     openMenu: "Abrir menu",
     myAccount: "Minha Conta",
@@ -227,114 +224,6 @@ const messages: Messages = {
     outOfStock: "{name} está sem estoque",
     limitedStock: "Apenas {count} unidades de {name} disponíveis",
     limitedUnits: "Apenas {count} unidades disponíveis",
-  },
-  fiscal: {
-    title: "Notas Fiscais",
-    subtitle: "Notas fiscais eletrônicas (NF-e / NFC-e)",
-    noInvoices: "Nenhuma nota fiscal encontrada",
-    emitInvoice: "Emitir Nota",
-    emitNfce: "Emitir NFC-e",
-    emitNfe: "Emitir NF-e",
-    syncPending: "Sincronizar Pendentes",
-    syncResult: "Sincronizadas {authorized} de {total} notas",
-    model: "Modelo",
-    series: "Série",
-    number: "Número",
-    accessKey: "Chave de Acesso",
-    recipient: "Destinatário",
-    issuedAt: "Emissão",
-    authorizedAt: "Autorização",
-    statusCode: "Código Status",
-    statusMessage: "Mensagem Status",
-    protocolNumber: "Protocolo",
-    contingency: "Contingência",
-    contingencyType: "Tipo Contingência",
-    contingencyReason: "Motivo",
-    downloadXml: "Baixar XML",
-    cancelInvoice: "Cancelar Nota",
-    cancelReason: "Motivo do cancelamento (mín. 15 caracteres)",
-    cancelSuccess: "Nota cancelada com sucesso",
-    cancelError: "Falha ao cancelar nota",
-    voidRange: "Inutilizar Faixa",
-    voidReason: "Motivo da inutilização (mín. 15 caracteres)",
-    startNumber: "Número Inicial",
-    endNumber: "Número Final",
-    voidSuccess: "Numeração inutilizada com sucesso",
-    voidError: "Falha ao inutilizar numeração",
-    emitSuccess: "Nota emitida com sucesso",
-    emitError: "Falha ao emitir nota",
-    emitContingency: "Nota emitida em contingência",
-    details: "Detalhes da Nota",
-    items: "Itens",
-    events: "Eventos",
-    xmlViewer: "XML",
-    authorized: "Autorizada",
-    rejected: "Rejeitada",
-    cancelled: "Cancelada",
-    denied: "Denegada",
-    pending: "Pendente",
-    voided: "Inutilizada",
-    nfe: "NF-e",
-    nfce: "NFC-e",
-    eventCancellation: "Cancelamento",
-    eventVoiding: "Inutilização",
-    sefazOnline: "SEFAZ Online",
-    sefazOffline: "SEFAZ Offline",
-    toggleNfce: "Emitir NFC-e",
-    searchPlaceholder: "Buscar notas fiscais...",
-    filterByStatus: "Status",
-    filterByModel: "Modelo",
-  },
-  fiscalSettings: {
-    title: "Configurações Fiscais",
-    subtitle: "Configure os dados fiscais da empresa e o certificado digital",
-    companyInfo: "Dados da Empresa",
-    companyName: "Razão Social",
-    tradeName: "Nome Fantasia",
-    taxId: "CNPJ",
-    stateTaxId: "Inscrição Estadual (IE)",
-    taxRegime: "Regime Tributário",
-    taxRegimeSimples: "Simples Nacional",
-    taxRegimeSimplesExcess: "Simples Nacional (excesso)",
-    taxRegimeNormal: "Normal (Lucro Presumido/Real)",
-    address: "Endereço",
-    stateCode: "Estado (UF)",
-    cityCode: "Código Município (IBGE)",
-    cityName: "Nome do Município",
-    street: "Logradouro",
-    streetNumber: "Número",
-    district: "Bairro",
-    zipCode: "CEP",
-    addressComplement: "Complemento",
-    environment: "Ambiente",
-    production: "Produção",
-    homologation: "Homologação (Testes)",
-    numbering: "Numeração",
-    nfeSeries: "Série NF-e",
-    nfceSeries: "Série NFC-e",
-    nextNfeNumber: "Próximo Nº NF-e",
-    nextNfceNumber: "Próximo Nº NFC-e",
-    nfceSettings: "Configurações NFC-e",
-    cscId: "ID do CSC",
-    cscToken: "Token CSC",
-    certificate: "Certificado Digital",
-    certificateFile: "Arquivo do Certificado (.pfx)",
-    certificatePassword: "Senha do Certificado",
-    certificateUploaded: "Certificado carregado",
-    certificateNotUploaded: "Nenhum certificado enviado",
-    certificateExpiry: "Válido até",
-    certificateCommonName: "Emitido para",
-    testConnection: "Testar Conexão",
-    testSuccess: "Conexão com SEFAZ bem-sucedida",
-    testError: "Falha na conexão com SEFAZ",
-    defaults: "Campos Fiscais Padrão",
-    defaultNcm: "NCM Padrão",
-    defaultCfop: "CFOP Padrão",
-    defaultIcmsCst: "CST ICMS Padrão",
-    defaultPisCst: "CST PIS Padrão",
-    defaultCofinsCst: "CST COFINS Padrão",
-    saved: "Configurações fiscais salvas",
-    saveError: "Falha ao salvar configurações fiscais",
   },
   error: {
     somethingWentWrong: "Desculpe, algo deu errado",
