+++ packages/addon-fiscal/src/messages.ts
@@ -1,0 +1,236 @@
+/**
+ * Mensagens i18n do addon fiscal — mescladas (deep) nas mensagens do host
+ * pelo addon-kit. O namespace "nav" contribui os labels de menu.
+ */
+export const fiscalMessages = {
+	en: {
+		nav: {
+			invoices: "Invoices",
+			fiscalSettings: "Fiscal Settings",
+			fiscal: "Fiscal",
+		},
+		fiscal: {
+			title: "Invoices",
+			subtitle: "Electronic invoices (NF-e / NFC-e)",
+			noInvoices: "No invoices found",
+			emitInvoice: "Emit Invoice",
+			emitNfce: "Emit NFC-e",
+			emitNfe: "Emit NF-e",
+			syncPending: "Sync Pending",
+			syncResult: "Synced {authorized} of {total} invoices",
+			model: "Model",
+			series: "Series",
+			number: "Number",
+			accessKey: "Access Key",
+			recipient: "Recipient",
+			issuedAt: "Issued At",
+			authorizedAt: "Authorized At",
+			statusCode: "Status Code",
+			statusMessage: "Status Message",
+			protocolNumber: "Protocol",
+			contingency: "Contingency",
+			contingencyType: "Contingency Type",
+			contingencyReason: "Reason",
+			downloadXml: "Download XML",
+			cancelInvoice: "Cancel Invoice",
+			cancelReason: "Cancellation reason (min. 15 characters)",
+			cancelSuccess: "Invoice cancelled successfully",
+			cancelError: "Failed to cancel invoice",
+			voidRange: "Void Number Range",
+			voidReason: "Voiding reason (min. 15 characters)",
+			startNumber: "Start Number",
+			endNumber: "End Number",
+			voidSuccess: "Numbers voided successfully",
+			voidError: "Failed to void numbers",
+			emitSuccess: "Invoice emitted successfully",
+			emitError: "Failed to emit invoice",
+			emitContingency: "Invoice emitted in contingency mode",
+			details: "Invoice Details",
+			items: "Items",
+			events: "Events",
+			xmlViewer: "XML",
+			authorized: "Authorized",
+			rejected: "Rejected",
+			cancelled: "Cancelled",
+			denied: "Denied",
+			pending: "Pending",
+			voided: "Voided",
+			nfe: "NF-e",
+			nfce: "NFC-e",
+			eventCancellation: "Cancellation",
+			eventVoiding: "Voiding",
+			sefazOnline: "SEFAZ Online",
+			sefazOffline: "SEFAZ Offline",
+			toggleNfce: "Emit NFC-e",
+			searchPlaceholder: "Search invoices...",
+			filterByStatus: "Status",
+			filterByModel: "Model",
+		},
+		fiscalSettings: {
+			title: "Fiscal Settings",
+			subtitle: "Configure your company fiscal data and digital certificate",
+			companyInfo: "Company Information",
+			companyName: "Company Name (Legal)",
+			tradeName: "Trade Name",
+			taxId: "CNPJ",
+			stateTaxId: "State Tax ID (IE)",
+			taxRegime: "Tax Regime",
+			taxRegimeSimples: "Simples Nacional",
+			taxRegimeSimplesExcess: "Simples Nacional (excess)",
+			taxRegimeNormal: "Normal (Lucro Presumido/Real)",
+			address: "Address",
+			stateCode: "State (UF)",
+			cityCode: "City Code (IBGE)",
+			cityName: "City Name",
+			street: "Street",
+			streetNumber: "Number",
+			district: "District",
+			zipCode: "ZIP Code (CEP)",
+			addressComplement: "Complement",
+			environment: "Environment",
+			production: "Production",
+			homologation: "Homologation (Testing)",
+			numbering: "Numbering",
+			nfeSeries: "NF-e Series",
+			nfceSeries: "NFC-e Series",
+			nextNfeNumber: "Next NF-e Number",
+			nextNfceNumber: "Next NFC-e Number",
+			nfceSettings: "NFC-e Settings",
+			cscId: "CSC ID",
+			cscToken: "CSC Token",
+			certificate: "Digital Certificate",
+			certificateFile: "Certificate File (.pfx)",
+			certificatePassword: "Certificate Password",
+			certificateUploaded: "Certificate loaded",
+			certificateNotUploaded: "No certificate uploaded",
+			certificateExpiry: "Valid until",
+			certificateCommonName: "Issued to",
+			testConnection: "Test Connection",
+			testSuccess: "SEFAZ connection successful",
+			testError: "SEFAZ connection failed",
+			defaults: "Default Fiscal Fields",
+			defaultNcm: "Default NCM",
+			defaultCfop: "Default CFOP",
+			defaultIcmsCst: "Default ICMS CST",
+			defaultPisCst: "Default PIS CST",
+			defaultCofinsCst: "Default COFINS CST",
+			saved: "Fiscal settings saved",
+			saveError: "Failed to save fiscal settings",
+		},
+	},
+	"pt-BR": {
+		nav: {
+			invoices: "Notas Fiscais",
+			fiscalSettings: "Configurações Fiscais",
+			fiscal: "Fiscal",
+		},
+		fiscal: {
+			title: "Notas Fiscais",
+			subtitle: "Notas fiscais eletrônicas (NF-e / NFC-e)",
+			noInvoices: "Nenhuma nota fiscal encontrada",
+			emitInvoice: "Emitir Nota",
+			emitNfce: "Emitir NFC-e",
+			emitNfe: "Emitir NF-e",
+			syncPending: "Sincronizar Pendentes",
+			syncResult: "Sincronizadas {authorized} de {total} notas",
+			model: "Modelo",
+			series: "Série",
+			number: "Número",
+			accessKey: "Chave de Acesso",
+			recipient: "Destinatário",
+			issuedAt: "Emissão",
+			authorizedAt: "Autorização",
+			statusCode: "Código Status",
+			statusMessage: "Mensagem Status",
+			protocolNumber: "Protocolo",
+			contingency: "Contingência",
+			contingencyType: "Tipo Contingência",
+			contingencyReason: "Motivo",
+			downloadXml: "Baixar XML",
+			cancelInvoice: "Cancelar Nota",
+			cancelReason: "Motivo do cancelamento (mín. 15 caracteres)",
+			cancelSuccess: "Nota cancelada com sucesso",
+			cancelError: "Falha ao cancelar nota",
+			voidRange: "Inutilizar Faixa",
+			voidReason: "Motivo da inutilização (mín. 15 caracteres)",
+			startNumber: "Número Inicial",
+			endNumber: "Número Final",
+			voidSuccess: "Numeração inutilizada com sucesso",
+			voidError: "Falha ao inutilizar numeração",
+			emitSuccess: "Nota emitida com sucesso",
+			emitError: "Falha ao emitir nota",
+			emitContingency: "Nota emitida em contingência",
+			details: "Detalhes da Nota",
+			items: "Itens",
+			events: "Eventos",
+			xmlViewer: "XML",
+			authorized: "Autorizada",
+			rejected: "Rejeitada",
+			cancelled: "Cancelada",
+			denied: "Denegada",
+			pending: "Pendente",
+			voided: "Inutilizada",
+			nfe: "NF-e",
+			nfce: "NFC-e",
+			eventCancellation: "Cancelamento",
+			eventVoiding: "Inutilização",
+			sefazOnline: "SEFAZ Online",
+			sefazOffline: "SEFAZ Offline",
+			toggleNfce: "Emitir NFC-e",
+			searchPlaceholder: "Buscar notas fiscais...",
+			filterByStatus: "Status",
+			filterByModel: "Modelo",
+		},
+		fiscalSettings: {
+			title: "Configurações Fiscais",
+			subtitle: "Configure os dados fiscais da empresa e o certificado digital",
+			companyInfo: "Dados da Empresa",
+			companyName: "Razão Social",
+			tradeName: "Nome Fantasia",
+			taxId: "CNPJ",
+			stateTaxId: "Inscrição Estadual (IE)",
+			taxRegime: "Regime Tributário",
+			taxRegimeSimples: "Simples Nacional",
+			taxRegimeSimplesExcess: "Simples Nacional (excesso)",
+			taxRegimeNormal: "Normal (Lucro Presumido/Real)",
+			address: "Endereço",
+			stateCode: "Estado (UF)",
+			cityCode: "Código Município (IBGE)",
+			cityName: "Nome do Município",
+			street: "Logradouro",
+			streetNumber: "Número",
+			district: "Bairro",
+			zipCode: "CEP",
+			addressComplement: "Complemento",
+			environment: "Ambiente",
+			production: "Produção",
+			homologation: "Homologação (Testes)",
+			numbering: "Numeração",
+			nfeSeries: "Série NF-e",
+			nfceSeries: "Série NFC-e",
+			nextNfeNumber: "Próximo Nº NF-e",
+			nextNfceNumber: "Próximo Nº NFC-e",
+			nfceSettings: "Configurações NFC-e",
+			cscId: "ID do CSC",
+			cscToken: "Token CSC",
+			certificate: "Certificado Digital",
+			certificateFile: "Arquivo do Certificado (.pfx)",
+			certificatePassword: "Senha do Certificado",
+			certificateUploaded: "Certificado carregado",
+			certificateNotUploaded: "Nenhum certificado enviado",
+			certificateExpiry: "Válido até",
+			certificateCommonName: "Emitido para",
+			testConnection: "Testar Conexão",
+			testSuccess: "Conexão com SEFAZ bem-sucedida",
+			testError: "Falha na conexão com SEFAZ",
+			defaults: "Campos Fiscais Padrão",
+			defaultNcm: "NCM Padrão",
+			defaultCfop: "CFOP Padrão",
+			defaultIcmsCst: "CST ICMS Padrão",
+			defaultPisCst: "CST PIS Padrão",
+			defaultCofinsCst: "CST COFINS Padrão",
+			saved: "Configurações fiscais salvas",
+			saveError: "Falha ao salvar configurações fiscais",
+		},
+	},
+};
