+++ packages/addon-fiscal/src/routers/cities.ts
@@ -1,0 +1,54 @@
+import { protectedProcedure, router } from "@finopenpos/api";
+import { lookupCep } from "@finopenpos/fiscal/cep-lookup";
+import { eq } from "drizzle-orm";
+import { z } from "zod/v4";
+import { db } from "../context";
+import { cities } from "../schema";
+
+const cepResultSchema = z.object({
+	zip_code: z.string(),
+	street: z.string(),
+	district: z.string(),
+	city_name: z.string(),
+	city_code: z.string(),
+	state_code: z.string(),
+});
+
+export const citiesRouter = router({
+	listByState: protectedProcedure
+		.meta({
+			openapi: {
+				method: "GET",
+				path: "/cities",
+				tags: ["Cities"],
+				summary: "List cities by state",
+			},
+		})
+		.input(z.object({ state_code: z.string().length(2) }))
+		.output(
+			z.array(
+				z.object({ id: z.number(), name: z.string(), state_code: z.string() }),
+			),
+		)
+		.query(async ({ input }) => {
+			return db
+				.select()
+				.from(cities)
+				.where(eq(cities.state_code, input.state_code.toUpperCase()));
+		}),
+
+	lookupCep: protectedProcedure
+		.meta({
+			openapi: {
+				method: "GET",
+				path: "/cities/cep/{cep}",
+				tags: ["Cities"],
+				summary: "Lookup address by CEP",
+			},
+		})
+		.input(z.object({ cep: z.string().min(8).max(9) }))
+		.output(cepResultSchema.nullable())
+		.query(async ({ input }) => {
+			return lookupCep(input.cep);
+		}),
+});
