+++ packages/addon-fiscal/tsconfig.json
@@ -1,0 +1,16 @@
+{
+	"extends": "@finopenpos/config/tsconfig.base.json",
+	"compilerOptions": {
+		"declaration": true,
+		"declarationMap": true,
+		"sourceMap": true,
+		"outDir": "dist",
+		"composite": true,
+		"jsx": "react-jsx",
+		"lib": ["ESNext", "DOM", "DOM.Iterable"],
+		// Código herdado do app web (que usa flags mais lenientes que o base)
+		"noUncheckedIndexedAccess": false,
+		"noUnusedLocals": false,
+		"noUnusedParameters": false
+	}
+}
