+++ packages/addon-kit/package.json
@@ -1,0 +1,37 @@
+{
+	"name": "@finopenpos/addon-kit",
+	"version": "0.0.0",
+	"private": true,
+	"type": "module",
+	"exports": {
+		".": {
+			"default": "./src/index.ts"
+		},
+		"./ui": {
+			"default": "./src/ui.tsx"
+		},
+		"./*": {
+			"default": "./src/*.ts"
+		}
+	},
+	"scripts": {
+		"check-types": "tsc --noEmit",
+		"test": "bun test src"
+	},
+	"dependencies": {
+		"@finopenpos/event-sourcing": "workspace:*",
+		"drizzle-orm": "^0.45.1"
+	},
+	"devDependencies": {
+		"@finopenpos/config": "workspace:*",
+		"@trpc/server": "catalog:",
+		"@types/bun": "^1.3.9",
+		"@types/react": "catalog:",
+		"react": "catalog:",
+		"typescript": "^5"
+	},
+	"peerDependencies": {
+		"@trpc/server": "catalog:",
+		"react": "catalog:"
+	}
+}
