+++ packages/addon-kit/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./registry";
+export * from "./types";
