+++ packages/addon-kit/tsconfig.json
@@ -1,0 +1,12 @@
+{
+	"extends": "@finopenpos/config/tsconfig.base.json",
+	"compilerOptions": {
+		"declaration": true,
+		"declarationMap": true,
+		"sourceMap": true,
+		"outDir": "dist",
+		"composite": true,
+		"jsx": "react-jsx",
+		"types": ["bun"]
+	}
+}
